% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BakariMtiliBCI-package.R
\docType{package}
\name{BakariMtiliBCI-package}
\alias{BakariMtiliBCI-package}
\alias{BakariMtiliBCI}
\title{BakariMtiliBCI}
\description{
\code{BakariMtiliSeries03} is a package developed as part of reproducible programming in R.
}
\details{
The package contains simple biodiversity data and simulation models.
}
\author{
Bakari Mtili <2500251M@student.gla.ac.uk> (maintainer)
}
