% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_data.R
\docType{data}
\name{bci_2010}
\alias{bci_2010}
\title{BCI census data 2010}
\format{
large matrix with 373750 elements (3.1MB)
}
\source{
\url{https://doi.org/10.5479/data.bci.20130603}
}
\usage{
bci_2010
}
\description{
The data from the BCI data set from 2010 saved as counts in a matrix.
}
\keyword{datasets}
