% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_data.R
\docType{data}
\name{bci_quadrats}
\alias{bci_quadrats}
\title{BCI Quadrat data}
\format{
1250 obs of 6 variables: QuadratName, row, col, x, y, Quadrat
}
\source{
\url{https://doi.org/10.5479/data.bci.20130603}
}
\usage{
bci_quadrats
}
\description{
The metadata of the quadrats
}
\keyword{datasets}
