% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_data.R
\docType{data}
\name{bci_taxa}
\alias{bci_taxa}
\title{BCI taxonomic data}
\format{
299 obs of 3 variables: GenusSpecies, Genus and Family
}
\source{
\url{https://doi.org/10.5479/data.bci.20130603}
}
\usage{
bci_taxa
}
\description{
The taxonomic data
}
\keyword{datasets}
