% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BakariMtiliBCI-package.R
\docType{package}
\name{BakariMtiliBCI-package}
\alias{BakariMtiliBCI-package}
\alias{BakariMtiliBCI}
\title{Barro Colorado Island data package}
\description{
This package contains BCI data and a number functions which will be used in simple analyses and simulations.
}
\author{
Bakari Mtili <2500251M@student.gla.ac.uk> (maintainer)
}
